backend/requirements.txt
Django
django-cors-headers
djangorestframework
psycopg2-binary==2.9.1